Moved ActionsPage role title formatting to a separate hook
